AI: Extract shared builder behavior into WP_AI_Client_Builder base class.
WP_AI_Client_Prompt_Builder and WP_AI_Client_Embedding_Builder shared
nearly all of their machinery: the snake_case __call() proxying, the
WP_Error error-state handling, exception-to-WP_Error conversion, and the
default request timeout setup. Move that into an abstract
WP_AI_Client_Builder base class, parameterized via abstract methods for
the SDK builder instantiation, error code prefix, prevent filter, and
generating/support-check method maps.

No behavior changes: error codes, filter names, and _doing_it_wrong
notices (which continue to report the concrete class name) are identical
to before.
